Move source files into new `src` directory
