+++ src/common-audio.js
