+++ src/common-polyfill.js
