+++ src/eq.html
