+++ src/eq.js
