+++ src/eqproc.js
