+++ src/qds.js
