+++ src/qdsproc.js
